kullanılan değişken tipleri ile ilgili açıklama yaptım.

+++ VariablesApp/Program.cs
@@ -10,24 +10,24 @@ namespace VariablesApp
     {
         static void Main(string[] args)
         {
-            Double tc ;
+            long tc ;
             Console.WriteLine("Lütfen asagıdaki bilgileri giriniz : ");
             Console.Write("T.C. Kimlik Numarasi : ");
-            tc = Convert.ToDouble(Console.ReadLine());
+            tc = Convert.ToInt64(Console.ReadLine()); 
             Console.Write("Adi : ");
             string ad = Console.ReadLine();
             Console.Write("Soyadi : ");
             string soyadi = Console.ReadLine();
             Console.Write("Telefon Numarasi : ");
-            double tel = Convert.ToDouble(Console.ReadLine());
+            double tel = Convert.ToDouble(Console.ReadLine()); // bu değişkende double kullnamamın sebebi int değişkenin alabileceğinden daha büyük bir sayı kullanıcak olmam
             Console.Write("Yaş : ");
             int yaş =Convert.ToInt32(Console.ReadLine());
             Console.Write("İlk Ürünün Fiyati : ");
-            double ilk = Convert.ToDouble(Console.ReadLine());
+            double ilk = Convert.ToDouble(Console.ReadLine()); // bu değişkende double kullnamamın sebebi ondalıklı bir sayı  kullanılabilir
             Console.Write("İkinci Ürünün Fiyati : ");
-            double ikinci = Convert.ToDouble(Console.ReadLine());
-            double toplam = ikinci + ilk;
-            double pp = toplam /10;
+            double ikinci = Convert.ToDouble(Console.ReadLine()); // bu değişkende double kullnamamın sebebi ondalıklı bir sayı kullanılabilir
+            double toplam = ikinci + ilk; // işlemin içinde kullanılan deişkenlerin double olması sebebi ile
+            double pp = toplam /10; // işlemin içinde kullanılan deişkenlerin double olması sebebi ile
             Console.WriteLine("--------------------------------------------");
             Console.WriteLine($"{tc} Tc Numarali {ad} isimli kisi için kayit olusturulmustur.");
             Console.WriteLine($"{tel} telefon numarasina bildirim mesaji gönderilmiştir.");
